refactoring - modify app name to lowercase & users -> user & Account -> accounts
