+++ .gitignore
@@ -6,21 +6,21 @@
 FD_backend/db.sqlite3
 FD_backend/db.sqlite3-journal
 FD_backend/FD_backend/__pycache__/
-FD_backend/Archive/__pycache__/
-FD_backend/users/__pycache__/
-FD_backend/Account/__pycache__/
-FD_backend/Restaurant/__pycache__/
-FD_backend/Review/__pycache__/
-FD_backend/Tag/__pycache__/
+FD_backend/archive/__pycache__/
+FD_backend/user/__pycache__/
+FD_backend/accounts/__pycache__/
+FD_backend/restaurant/__pycache__/
+FD_backend/review/__pycache__/
+FD_backend/tag/__pycache__/
 FD_backend/api/__pycache__/
 
-FD_backend/Account/migrations/
-FD_backend/Archive/migrations/
-FD_backend/Ranking/migrations/
-FD_backend/Restaurant/migrations/
-FD_backend/Review/migrations/
-FD_backend/Tag/migrations/
-FD_backend/users/migrations/
+FD_backend/accounts/migrations/
+FD_backend/archive/migrations/
+FD_backend/ranking/migrations/
+FD_backend/restaurant/migrations/
+FD_backend/review/migrations/
+FD_backend/tag/migrations/
+FD_backend/user/migrations/
 
 FD_backend/.coverage/
 FD_backend/photo/
