+++ FD_backend/Archive/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ArchiveConfig(AppConfig):
-    name = 'Archive'
+    name = 'archive'
