+++ FD_backend/Restaurant/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class RestaurantConfig(AppConfig):
-    name = 'Restaurant'
+    name = 'restaurant'
