+++ FD_backend/Review/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ReviewConfig(AppConfig):
-    name = 'Review'
+    name = 'review'
