+++ FD_backend/Tag/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TagConfig(AppConfig):
-    name = 'Tag'
+    name = 'tag'
