+++ FD_backend/Tag/models.py
@@ -1,5 +1,3 @@
-# Tag/models.py
-
 from django.db import models
 
 # Create your models here.
