+++ FD_backend/accounts/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AccountsConfig(AppConfig):
+    name = 'accounts'
