+++ FD_backend/user/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UserConfig(AppConfig):
+    name = 'user'
