+++ FD_backend/user/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 from . import models
-from Archive.models import Archive
+from archive.models import Archive
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
